bump v0.1.0
Commit for npm publish. The commit #78c8ec0 that resolves the ViewPropTypes issue has not been reflected in the npm package, so I'm releasing the npm package as version 0.1.0 to include it. Recent work done to address the ViewPropTypes issue has been determined to be unrelated and has been removed.